Funcionalidades básicas completadas

+++ controllers/polls.controller.js
@@ -2,24 +2,49 @@ const {Poll} = require("../models/Poll");
 
 const getPoll = async (req, res) => {
     const poll = await Poll.findById(req.params.id);
-    res.json(poll);
+
+    const pollData = {
+        createdBy: poll.createdBy,
+        status: poll.status,
+        question: poll.question,
+        description: poll.description,
+        options: poll.options,
+        createdAt: poll.createdAt,
+        endDate: poll.endDate
+    };
+
+    res.json(pollData);
 }
 
-const getDashboard = async (req, res) => {
-    if (req.user.adminRole === true) {
-        try {
-            const allPolls = await Poll.find();
-            res.json(allPolls); 
-        } catch (error) {
-            console.log(error);
-            res.status(400).send("Unexpected error");
-        }
-    }
-    else {
-        res.status(401).send("Unauthorized to get dashboard");
+const votePoll = async (req, res) => {
+    const pollId = await Poll.findById(req.params.id);
+
+    if (!pollId) {
+        res.status(404).send("Poll not found");
         return;
     }
-    
+
+    if (pollId.status !== "Open") {
+        res.status(400).send("Poll is closed");
+        return;
+    }
+
+    const existingVote = await Poll.findOne({ _id: pollId, "votes.userId": req.user._id });
+    if (existingVote) {
+        res.status(400).send("You have already voted in this poll.");
+        return;
+    }
+
+    const userChoice = req.body.userChoice;
+    const userVote = { userId: req.user._id, vote: userChoice };
+
+    try {
+        const insertNewVotetoPoll = await Poll.findByIdAndUpdate(pollId, {$push: { votes: userVote }});
+        res.send("Voted successfully");
+    } catch (error) {
+        console.log(error);
+        res.status(400).send("Unexpected error");
+    }
 }
 
 const createPoll = async (req, res) => {
@@ -30,7 +55,7 @@ const createPoll = async (req, res) => {
 
     try {
         const options = req.body.options.map((option, index) => ({
-            id: index + 1, // o puedes utilizar un UUID si lo prefieres
+            id: index + 1, 
             value: option,
         }));
 
@@ -44,18 +69,19 @@ const createPoll = async (req, res) => {
 }
 
 const updatePoll = async (req, res) => {
-    const pollId = await Poll.findById(req.params.id);
-    const userChoice = req.body.userChoice;
-
-    const userVote = { userId: req.user._id, vote: userChoice };
-
-    const existingVote = await Poll.findOne({ _id: pollId, "votes.userId": req.user._id });
-    if (existingVote) {
-        return res.status(400).send("You have already voted in this poll.");
+    if (req.user.adminRole === false) {
+        res.status(401).send("Unauthorized to update polls");
+        return;
     }
 
-    const updatedPoll = await Poll.findByIdAndUpdate(pollId, {$push: { votes: userVote }});
-    res.json(updatedPoll);
+    try {
+        const pollId = await Poll.findById(req.params.id);
+        const updatedPoll = await Poll.findByIdAndUpdate(pollId, req.body);
+        res.json(updatedPoll);
+    } catch (error) {
+        console.log(error);
+        res.status(400).send("Unexpected error");
+    }
 }
 
 const closePoll = async (req, res) => {
@@ -66,8 +92,8 @@ const closePoll = async (req, res) => {
 
     try {
         const pollId = await Poll.findById(req.params.id);
-        const updatedPoll = await Poll.findByIdAndUpdate(pollId, { status: "Closed"});
-        res.json(updatedPoll);
+        const updatedPoll = await Poll.findByIdAndUpdate(pollId, { status: "Closed", endDate: new Date() });
+        res.send("Poll closed");
     } catch (error) {
         console.log(error);
         res.status(400).send("Unexpected error");
@@ -75,6 +101,11 @@ const closePoll = async (req, res) => {
 }
 
 const deletePoll = async (req, res) => {
+    if (req.user.adminRole === false) {
+        res.status(401).send("Unauthorized to remove polls");
+        return;
+    }
+
     const deletedPoll = await Poll.findByIdAndDelete({pollId: req.params.id});
     res.json(deletedPoll);
 }
@@ -97,4 +128,21 @@ const getPollResults = async (req, res) => {
     res.json(pollResults);
 }
 
-module.exports = { getPoll, getDashboard, createPoll, updatePoll, closePoll, deletePoll, getPollResults };
+const getDashboard = async (req, res) => {
+    if (req.user.adminRole === true) {
+        try {
+            const allPolls = await Poll.find();
+            res.json(allPolls); 
+        } catch (error) {
+            console.log(error);
+            res.status(400).send("Unexpected error");
+        }
+    }
+    else {
+        res.status(401).send("Unauthorized to get dashboard");
+        return;
+    }
+    
+}
+
+module.exports = { getPoll, votePoll, createPoll, updatePoll, closePoll, deletePoll, getPollResults, getDashboard };
